requirements.txt
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
opencv-python>=4.8.0
albumentations>=1.3.0
timm>=0.9.0
pytorch-lightning>=2.0.0
wandb>=0.15.0
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0
pyyaml>=6.0
hydra-core>=1.2.0
omegaconf>=2.2.0
tqdm>=4.65.0
matplotlib>=3.7.0
seaborn>=0.12.0
grad-cam>=1.4.0
fastapi>=0.100.0
uvicorn>=0.23.0
